Add an option to skip setting CMAKE_DEBUG_POSTFIX
In our host CMakeLists.txt, we were doing add_subdirectory(expat). The
expat CMakeLists.txt was setting CMAKE_DEBUG_POSTFIX, which caused all
of the other debug DLLs in our host CMakeLists.txt to be postfixed with 'd'.

We don't want any of our DLLs postfixed with 'd', so this adds an option
to disable that.